requirements.txt
flet>=0.21.0
Pillow>=10.0.0
reportlab>=4.0.0
